add local_timezone_string_provider
